delete WorkspaceResourceMonitor and its runners
