+++ core/src/main/scala/org/broadinstitute/dsde/rawls/monitor/BootMonitors.scala
@@ -6,46 +6,20 @@ import cats.effect.IO
 import com.typesafe.config.{Config, ConfigRenderOptions}
 import com.typesafe.scalalogging.LazyLogging
 import net.ceedubs.ficus.Ficus.{optionValueReader, toFicusConfig}
-import org.broadinstitute.dsde.rawls.billing.{BillingProfileManagerDAO, BillingProjectDeletion, BillingRepository}
+import org.broadinstitute.dsde.rawls.billing.BillingProfileManagerDAO
 import org.broadinstitute.dsde.rawls.config.{FastPassConfig, RawlsConfigManager}
-import org.broadinstitute.dsde.rawls.coordination.{
-  CoordinatedDataSourceAccess,
-  CoordinatedDataSourceActor,
-  DataSourceAccess,
-  UncoordinatedDataSourceAccess
-}
+import org.broadinstitute.dsde.rawls.coordination.{CoordinatedDataSourceAccess, CoordinatedDataSourceActor, DataSourceAccess, UncoordinatedDataSourceAccess}
 import org.broadinstitute.dsde.rawls.dataaccess._
 import org.broadinstitute.dsde.rawls.dataaccess.drs.DrsResolver
-import org.broadinstitute.dsde.rawls.dataaccess.leonardo.LeonardoService
-import org.broadinstitute.dsde.rawls.dataaccess.slick.WorkspaceManagerResourceMonitorRecord.JobType
 import org.broadinstitute.dsde.rawls.dataaccess.workspacemanager.WorkspaceManagerDAO
 import org.broadinstitute.dsde.rawls.entities.EntityService
 import org.broadinstitute.dsde.rawls.fastpass.FastPassMonitor
 import org.broadinstitute.dsde.rawls.google.GooglePubSubDAO
-import org.broadinstitute.dsde.rawls.jobexec.{
-  MethodConfigResolver,
-  SubmissionMonitorConfig,
-  SubmissionSupervisor,
-  WorkflowSubmissionActor
-}
+import org.broadinstitute.dsde.rawls.jobexec.{MethodConfigResolver, SubmissionMonitorConfig, SubmissionSupervisor, WorkflowSubmissionActor}
 import org.broadinstitute.dsde.rawls.metrics.BardService
-import org.broadinstitute.dsde.rawls.model.{
-  CromwellBackend,
-  RawlsRequestContext,
-  WorkflowStatuses,
-  WorkspaceCloudPlatform
-}
+import org.broadinstitute.dsde.rawls.model.{CromwellBackend, RawlsRequestContext, WorkflowStatuses, WorkspaceCloudPlatform}
 import org.broadinstitute.dsde.rawls.monitor.AvroUpsertMonitorSupervisor.AvroUpsertMonitorConfig
 import org.broadinstitute.dsde.rawls.monitor.migration.MultiregionalBucketMigrationActor
-import org.broadinstitute.dsde.rawls.monitor.workspace.WorkspaceResourceMonitor
-import org.broadinstitute.dsde.rawls.monitor.workspace.runners.clone.WorkspaceCloningRunner
-import org.broadinstitute.dsde.rawls.monitor.workspace.runners.deletion.WorkspaceDeletionRunner
-import org.broadinstitute.dsde.rawls.monitor.workspace.runners.deletion.actions.WsmDeletionAction
-import org.broadinstitute.dsde.rawls.monitor.workspace.runners.{
-  BPMBillingProjectDeleteRunner,
-  CloneWorkspaceContainerRunner,
-  LandingZoneCreationStatusRunner
-}
 import org.broadinstitute.dsde.rawls.util
 import org.broadinstitute.dsde.rawls.workspace.{WorkspaceRepository, WorkspaceService, WorkspaceSettingRepository}
 import org.broadinstitute.dsde.workbench.dataaccess.NotificationDAO
@@ -229,18 +203,6 @@ object BootMonitors extends LazyLogging {
       util.toScalaDuration(cloneWorkspaceFileTransferMonitorConfigRoot.getDuration("initialDelay"))
     )
     startCloneWorkspaceFileTransferMonitor(system, cloneWorkspaceFileTransferMonitorConfig, slickDataSource, gcsDAO)
-
-    startWorkspaceResourceMonitor(
-      system,
-      appConfigManager.conf,
-      slickDataSource,
-      samDAO,
-      workspaceManagerDAO,
-      billingProfileManagerDAO,
-      gcsDAO,
-      leonardoDAO,
-      workspaceRepository
-    )
 
   }
 
@@ -457,63 +419,6 @@ object BootMonitors extends LazyLogging {
         dataSource
       )
     )
-
-  private def startWorkspaceResourceMonitor(
-    system: ActorSystem,
-    config: Config,
-    dataSource: SlickDataSource,
-    samDAO: SamDAO,
-    workspaceManagerDAO: WorkspaceManagerDAO,
-    billingProfileManagerDAO: BillingProfileManagerDAO,
-    gcsDAO: GoogleServicesDAO,
-    leonardoDAO: LeonardoDAO,
-    workspaceRepository: WorkspaceRepository
-  ) = {
-    val billingRepo = new BillingRepository(dataSource)
-
-    val leoService = new LeonardoService(leonardoDAO)(system)
-    val wsmDeletionAction = new WsmDeletionAction(workspaceManagerDAO)(system)
-    val monitorRecordDao = WorkspaceManagerResourceMonitorRecordDao(dataSource)
-    val workspaceDeletionRunner = new WorkspaceDeletionRunner(samDAO,
-                                                              workspaceManagerDAO,
-                                                              workspaceRepository,
-                                                              leoService,
-                                                              wsmDeletionAction,
-                                                              gcsDAO,
-                                                              monitorRecordDao
-    )
-    val workspaceCloneRunner = new WorkspaceCloningRunner(
-      samDAO,
-      gcsDAO,
-      leonardoDAO,
-      workspaceManagerDAO,
-      monitorRecordDao,
-      workspaceRepository
-    )
-    system.actorOf(
-      WorkspaceResourceMonitor.props(
-        config,
-        dataSource,
-        Map(
-          JobType.WorkspaceDeleteInit -> workspaceDeletionRunner,
-          JobType.LeoAppDeletionPoll -> workspaceDeletionRunner,
-          JobType.LeoRuntimeDeletionPoll -> workspaceDeletionRunner,
-          JobType.WSMWorkspaceDeletionPoll -> workspaceDeletionRunner,
-          JobType.AzureLandingZoneResult ->
-            new LandingZoneCreationStatusRunner(samDAO, workspaceManagerDAO, billingRepo, gcsDAO),
-          JobType.CloneWorkspaceContainerResult ->
-            new CloneWorkspaceContainerRunner(samDAO, workspaceManagerDAO, dataSource, gcsDAO),
-          JobType.BpmBillingProjectDelete -> new BPMBillingProjectDeleteRunner(
-            samDAO,
-            gcsDAO,
-            workspaceManagerDAO,
-            billingRepo,
-            new BillingProjectDeletion(samDAO, billingRepo, billingProfileManagerDAO)
-          )
-        ) ++ JobType.cloneJobTypes.map(jobType => jobType -> workspaceCloneRunner).toMap
-      )
-    )
-  }
 
   private def startMultiregonalBucketMigrationActor(system: ActorSystem,
                                                     config: Config,
