--- core/src/main/scala/org/broadinstitute/dsde/rawls/monitor/workspace/runners/deletion/actions/WorkspaceDeletionActionFailureException.scala
@@ -1,5 +1,0 @@
-package org.broadinstitute.dsde.rawls.monitor.workspace.runners.deletion.actions
-
-import org.broadinstitute.dsde.rawls.RawlsException
-
-case class WorkspaceDeletionActionFailureException(message: String) extends RawlsException(message) {}
